refactor(modules): promote VQ (Codebook/VectorQuantizer/ResidualVQ) to braindecode.modules.quantization
